backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.0
django-cors-headers==4.3.1
reportlab==4.0.9
Pillow>=10.0.0
python-dotenv==1.0.0
